refactor: Popout FloatButton component
